inverse of half

+++ src/diffusers/pipelines/stable_diffusion/pipeline_stable_diffusion.py
@@ -1057,12 +1057,12 @@ class StableDiffusionPipeline(
                     if riemann:
                         print('riemann')
                         def metric_tensor(score):
-                            score_term_1 = score.permute(0, 2, 3, 1).unsqueeze(-1)
-                            score_term_2 = score.permute(0, 2, 3, 1).unsqueeze(-2)
+                            score_term_1 = score.permute(0, 2, 3, 1).unsqueeze(-1).to(torch.float32)
+                            score_term_2 = score.permute(0, 2, 3, 1).unsqueeze(-2).to(torch.float32)
                             G = score_term_1 @ score_term_2
                             G = G.mean(dim=0)
                             G_inv = torch.linalg.inv(G)
-                            return G_inv
+                            return G_inv.to(torch.float16)
                         def mm(A, B):# A is 32 x 32 x 3 x 3 and B is bs x 3 x 32 x 32
                             # Use the same dtype as the input tensors to avoid dtype mismatch
                             target_dtype = B.dtype
